refactor: uploadBox -> photoUploadButton으로 수정
