+++ frontend/src/components/host/photoUploadButton/PhotoUploadButton.styles.ts
